fix(grumbles): remove unncessary classes, event listeners, methods, abstractions. Also revise for styleguide
