--- packages/electron-app/src/main/app/methods/index.ts
@@ -1,12 +1,0 @@
-// Copyright 2018-2019 @paritytech/substrate-light-ui authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-import createWindow from './createWindow';
-import setupAppListeners from './setupAppListeners';
-import setupParitySubstrate from './setupParitySubstrate';
-
-export {
-    createWindow,
-    setupAppListeners,
-    setupParitySubstrate
-};
